movie page response

+++ src/main/java/com/jupitters/movieApi/dto/MoviePageResponse.java
@@ -5,7 +5,7 @@ import java.util.List;
 public record MoviePageResponse(List<MovieDto> movieDtos,
                                 Integer pageNumber,
                                 Integer pageSize,
-                                Integer totalElements,
+                                Long totalElements,
                                 Integer totalPages,
                                 boolean isLast) {
